test: integrate unit tests existed in tests.rs to functional_tests.rs
