+++ .github/workflows/release.yml
@@ -68,8 +68,11 @@ jobs:
           body: |
             ## ${{ github.ref_name }}
 
-            - [Documentation](https://docs.rs/rsext4)
-            - [crates.io](https://crates.io/crates/rsext4)
+            - [Documentation](https://arceos.org/axfs_crates/)
+            - crates.io:
+              - [axfs_vfs](https://crates.io/crates/axfs_vfs)
+              - [axfs_ramfs](https://crates.io/crates/axfs_ramfs)
+              - [axfs_devfs](https://crates.io/crates/axfs_devfs)
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
