+++ axfs_devfs/src/lib.rs
@@ -29,9 +29,6 @@ mod null;
 mod urandom;
 mod zero;
 
-#[cfg(test)]
-mod tests;
-
 pub use self::dir::DirNode;
 pub use self::null::NullDev;
 pub use self::urandom::UrandomDev;
